+++ src/data/grl-data.h
@@ -97,7 +97,7 @@ void grl_data_set_int (GrlData *data, GrlKeyID key, gint intvalue);
 
 void grl_data_set_float (GrlData *data,
                          GrlKeyID key,
-                         gint floatvalue);
+                         gfloat floatvalue);
 
 const GValue *grl_data_get (GrlData *data, GrlKeyID key);
